cleanup quartz plugin
