--- apm-agent-plugins/apm-quartz-job-plugin/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
-    <modules>
-        <module>quartz-common</module>
-        <module>apm-quartz-1-plugin</module>
-        <module>apm-quartz-2-plugin</module>
-    </modules>
-
-    <parent>
-        <artifactId>apm-agent-plugins</artifactId>
-        <groupId>co.elastic.apm</groupId>
-        <version>1.38.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>apm-quartz-job-plugin</artifactId>
-    <name>${project.groupId}:${project.artifactId}</name>
-
-    <properties>
-        <apm-agent-parent.base.dir>${project.basedir}/../..</apm-agent-parent.base.dir>
-    </properties>
-
-</project>
