+++ apm-agent-plugins/apm-quartz/apm-quartz-plugin-1/src/main/java/co/elastic/apm/agent/quartzjob/Quartz1JobExecutionContextHandler.java
