+++ apm-agent-plugins/apm-quartz/apm-quartz-plugin-1/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
